fix(autocmds) : fix showing no task error when existed file is opened
harper-ls process is checked at VimEnter to execute once.
But VimEnter is invoked too late so it cannot check the number of
harper-ls process before the file is read. it invokes error

1) change autocmd event to BufReadPre / BufNewFile
2) check the the number of harper-ls
3) set `once` option to execute the autocmd at startup

+++ lua/jaehak/core/autocmds.lua
@@ -63,14 +63,18 @@ vim.api.nvim_create_autocmd({"VimLeave"}, {
 })
 
 -- Check redundant process and terminate at startup
-vim.api.nvim_create_autocmd({"VimEnter"}, {
+vim.api.nvim_create_autocmd({"BufReadPre", "BufNewFile"}, {
 	group = SystemCall,
 	callback = function ()
 		local nvim_process = GetProcessId('nvim.exe')
 		if #nvim_process <= 3 then
-			vim.fn.system('taskkill /F /IM ' .. 'harper-ls.exe')
+			local harper_ls_process = #GetProcessId('harper-ls.exe')
+			if harper_ls_process > 0 then
+				vim.fn.system('taskkill /F /IM ' .. 'harper-ls.exe')
+			end
 		end
-	end
+	end,
+	once = true,
 })
